+++ packages/@granit/cms-redirects/src/index.ts
@@ -1,0 +1,19 @@
+// ─── Types ───────────────────────────────────────────────────────────────────
+export type {
+  CreateRedirectRequest,
+  RedirectResolveResponse,
+  RedirectResponse,
+  UpdateRedirectRequest,
+} from './types/index.js';
+
+// ─── API — Public renderer ────────────────────────────────────────────────────
+export { resolveRedirect } from './api/redirects.js';
+
+// ─── API — Admin ─────────────────────────────────────────────────────────────
+export {
+  createRedirect,
+  deleteRedirect,
+  listRedirects,
+  updateRedirect,
+} from './api/redirects-admin.js';
+export type { ListRedirectsParams } from './api/redirects-admin.js';
