+++ packages/@granit/cms-seo/package.json
@@ -1,0 +1,17 @@
+{
+  "name": "@granit/cms-seo",
+  "description": "Granit CMS — SEO types, public effective-SEO client, SEO admin and SEO-AI API",
+  "version": "0.1.0",
+  "license": "Apache-2.0",
+  "type": "module",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "peerDependencies": {
+    "@granit/api-client": "workspace:*"
+  },
+  "devDependencies": {
+    "@granit/api-client": "workspace:*",
+    "@granit/testing": "workspace:*"
+  }
+}
