+++ packages/@granit/cms-seo/src/api/seo-ai.ts
